organize component logic
